Moved Cog to a new file for maintenance
